refactor(ts): extract generic fmt, delegate s3/ram/disk/redis/opfs
fmt was duplicated across 5 backends with identical wrapText/fmtText logic.
Add generic/fmt.ts (exported from core) and reduce each backend to a thin
wrapper injecting its stream, matching python's single generic.fmt.
