+++ typescript/packages/browser/src/commands/builtin/opfs/fmt.ts
@@ -13,13 +13,10 @@
 // ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
 import {
-  IOResult,
   ResourceName,
   command,
-  materialize,
-  readStdinAsync,
+  fmtGeneric,
   specOf,
-  type ByteSource,
   type CommandFnResult,
   type CommandOpts,
   type PathSpec,
@@ -27,61 +24,13 @@ import {
 import { stream as opfsStream } from '../../../core/opfs/stream.ts'
 import type { OPFSAccessor } from '../../../accessor/opfs.ts'
 
-const ENC = new TextEncoder()
-const DEC = new TextDecoder('utf-8', { fatal: false })
-
-function wrapText(text: string, width: number): string {
-  const words = text.split(/\s+/).filter((w) => w !== '')
-  if (words.length === 0) return ''
-  const lines: string[] = []
-  let current = ''
-  for (const w of words) {
-    if (current === '') {
-      current = w
-    } else if (current.length + 1 + w.length <= width) {
-      current = current + ' ' + w
-    } else {
-      lines.push(current)
-      current = w
-    }
-  }
-  if (current !== '') lines.push(current)
-  return lines.join('\n')
-}
-
-function fmtText(text: string, width: number): string {
-  const paragraphs = text.split('\n\n')
-  const formatted: string[] = []
-  for (const para of paragraphs) {
-    const trimmed = para.trim()
-    if (trimmed !== '') formatted.push(wrapText(trimmed, width))
-    else formatted.push('')
-  }
-  return formatted.join('\n\n') + '\n'
-}
-
 async function fmtCommand(
   accessor: OPFSAccessor,
   paths: PathSpec[],
-  texts: string[],
+  _texts: string[],
   opts: CommandOpts,
 ): Promise<CommandFnResult> {
-  const width = typeof opts.flags.w === 'string' ? Number.parseInt(opts.flags.w, 10) : 75
-  if (paths.length > 0) {
-    const parts: string[] = []
-    for (const p of paths) {
-      parts.push(DEC.decode(await materialize(opfsStream(accessor, p))))
-    }
-    const result: ByteSource = ENC.encode(fmtText(parts.join(''), width))
-    return [result, new IOResult()]
-  }
-  const stdinData = await readStdinAsync(opts.stdin)
-  if (stdinData === null) {
-    return [null, new IOResult({ exitCode: 1, stderr: ENC.encode('fmt: missing operand\n') })]
-  }
-  const text = DEC.decode(stdinData)
-  const result: ByteSource = ENC.encode(fmtText(text, width))
-  return [result, new IOResult()]
+  return fmtGeneric(paths, opts, (p) => opfsStream(accessor, p))
 }
 
 export const OPFS_FMT = command({
